release v2.3.2

+++ src/main/java/com/lokoko/domain/productReview/api/dto/response/ImageReviewResponse.java
@@ -12,7 +12,7 @@ public record ImageReviewResponse(
         @Schema(requiredMode = REQUIRED)
         String productName,
         @Schema(requiredMode = REQUIRED)
-        Integer likeCount,
+        Long likeCount,
         @Schema(requiredMode = REQUIRED)
         String url
 ) {
